feature: 동호회 내 게시글 조회 API #62

+++ src/main/java/com/hansol/tofu/board/domain/dto/BoardResponseDTO.java
@@ -16,7 +16,7 @@ public record BoardResponseDTO(
 	String memberName,
 	String title,
 	String content,
-	int commentCount,
+	Long commentCount,
 	ZonedDateTime createdAt,
 	ZonedDateTime updatedAt,
 	List<ClubPhotoResponseDTO> photoList
